Solving Notify Following Users Bug Adding Current User Active Orders Count Can Order By User Balance

+++ LowCost.Business/Services/ProductFollowingUsersService/Implementation/ProductFollowingUsersService.cs
@@ -2,6 +2,8 @@
 using LowCost.Domain.Models;
 using LowCost.Infrastructure.Helpers;
 using LowCost.Repo.UnitOfWork;
+using LowCost.Resources;
+using Microsoft.Extensions.Localization;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -12,10 +14,12 @@ namespace LowCost.Business.Services.ProductFollowingUsersService.Implementation
     public class ProductFollowingUsersService : IProductFollowingUsersService
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IStringLocalizer<SharedResource> _stringLocalizer;
 
-        public ProductFollowingUsersService(IUnitOfWork unitOfWork)
+        public ProductFollowingUsersService(IUnitOfWork unitOfWork, IStringLocalizer<SharedResource> stringLocalizer)
         {
             this._unitOfWork = unitOfWork;
+            this._stringLocalizer = stringLocalizer;
         }
 
 
@@ -25,16 +29,26 @@ namespace LowCost.Business.Services.ProductFollowingUsersService.Implementation
             // Get Current User Id
             var currentUserid = await _unitOfWork.CurrentUserRepository.GetCurrentUserId();
             // Adding User To Product Following Users
-            var following = new ProductFollowingUser
+            var following = await _unitOfWork.ProductFollowingUsersRepository
+                                 .FindElementAsync(follower => follower.Product_Id == product_Id && follower.User_Id == currentUserid);
+           
+            if(following == null)
             {
-                Product_Id = product_Id,
-                User_Id = currentUserid
-            };
-            await _unitOfWork.ProductFollowingUsersRepository.CreateAsync(following);
-            var result = await _unitOfWork.SaveAsync() > 0;
-            if(result)
+                following = new ProductFollowingUser
+                {
+                    Product_Id = product_Id,
+                    User_Id = currentUserid
+                };
+                await _unitOfWork.ProductFollowingUsersRepository.CreateAsync(following);
+                var result = await _unitOfWork.SaveAsync() > 0;
+                if (result)
+                {
+                    actionState.ExcuteSuccessfully = true;
+                }
+            }
+            else
             {
-                actionState.ExcuteSuccessfully = true;
+                actionState.ErrorMessages.Add(_stringLocalizer["You Already Follow This Product"]);
             }
             return actionState;
         }
@@ -48,12 +62,20 @@ namespace LowCost.Business.Services.ProductFollowingUsersService.Implementation
             var following = await _unitOfWork.ProductFollowingUsersRepository
                                  .FindElementAsync(follower => follower.Product_Id == product_Id && follower.User_Id == currentUserid);
            
-            _unitOfWork.ProductFollowingUsersRepository.Delete(following);
-            var result = await _unitOfWork.SaveAsync() > 0;
-            if (result)
+            if(following != null)
             {
-                actionState.ExcuteSuccessfully = true;
+                _unitOfWork.ProductFollowingUsersRepository.Delete(following);
+                var result = await _unitOfWork.SaveAsync() > 0;
+                if (result)
+                {
+                    actionState.ExcuteSuccessfully = true;
+                }
             }
+            else
+            {
+                actionState.ErrorMessages.Add(_stringLocalizer["You Are Not Follow This Product"]);
+            }
+
             return actionState;
         }
     }
